feat: Set application state values
In this commit, I set up the state properties of the home and away teams. This was done using
the useState hook available in the react library. A state of score and a state setter of set-
score was established for both the home and away team, with it's initial value initialized
to 10 and 5 for the home and away teams respectively.